pyproject.toml
[tool.poetry]
name = "repl_python3_Vortex-Bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
dnspython = "^2.1.0"
animec = "^0.3.3"
Flask = "^2.0.1"
utils = "^1.0.1"
pymongo = "^3.12.0"
psutil = "^5.8.0"
messages = "^0.8.0"
aiohttp = "^3.8.1"
disnake-together = "^1.0.3"
disrank2 = "^0.0.4"
Pillow = "^8.4.0"
jishaku = "^2.3.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
